+++ R/pattern.R
@@ -52,7 +52,7 @@ all_patterns = list(
     inline.comment = '^###[.].*')
 )
 
-.sep.label = '^#+\\s*(@knitr|----+)(.*?)-*\\s*$'  # pattern for code chunks in an R script
+.sep.label = '^(#|--)+\\s*(@knitr|----+)(.*?)-*\\s*$'  # pattern for code chunks in an R script
 
 # initial pattern list
 .pat.init = list(